[AOT Refactor] functional utils
[ghstack-poisoned]
