+++ torch/_functorch/aot_autograd.py
@@ -31,7 +31,7 @@ from torch._subclasses.functional_tensor import FunctionalTensor, FunctionalTens
 from torch.fx import Interpreter
 from torch.fx.experimental.proxy_tensor import is_sym_node
 from torch.fx.experimental.symbolic_shapes import (
-    ShapeEnv, is_concrete_int, fx_placeholder_vals, definitely_true, definitely_false, sym_eq
+    ShapeEnv, is_concrete_int, fx_placeholder_vals, definitely_false, sym_eq
 )
 from torch.multiprocessing.reductions import StorageWeakRef
 from torch.nn.utils import stateless
@@ -53,6 +53,18 @@ from ._aot_autograd.logging_utils import (  # noqa: F401
     get_graph_being_compiled,
     get_aot_compilation_context,
     track_graph_compiling,
+)
+from ._aot_autograd.functional_utils import (
+    is_functional,
+    to_functional,
+    from_functional,
+    sync_functional_tensor,
+    has_metadata_mutation,
+    are_all_mutations_hidden_from_autograd,
+    gen_alias_from_base,
+    has_same_metadata,
+    was_tensor_updated,
+    was_tensor_metadata_updated,
 )
 
 zip = strict_zip
@@ -703,192 +715,6 @@ class TensorAlias:
     alias: torch.Tensor
 
 
-def has_same_metadata(t1, t2):
-    return (
-        definitely_true(sym_eq(t1.size(), t2.size()))
-        and definitely_true(sym_eq(t1.stride(), t2.stride()))
-        and definitely_true(t1.storage_offset() == t2.storage_offset())
-        and t1.is_conj() == t2.is_conj()
-        and t1.is_neg() == t2.is_neg()
-    )
-
-
-def gen_alias_from_base(aliased_base_tensor, target_meta_tensor, target_requires_grad):
-    # Try to do view-replay if possible.
-    # fall back to .as_strided() if we can't.
-    if target_meta_tensor._base is not None:
-        # The base that we want to replay our view off of might have a different shape than the view's original base.
-        b = target_meta_tensor._base
-        abt = aliased_base_tensor
-        # Don't unnecessarily call as_strided if nothing changed; as_strided's
-        # backward is poorly implemented and slow
-        if abt is not b and (
-            abt.size() != b.size() or
-            abt.stride() != b.stride() or
-            abt.storage_offset() != b.storage_offset()
-        ):
-            reshaped_base_tensor = aliased_base_tensor.as_strided(
-                b.size(), b.stride(), b.storage_offset()
-            )
-        else:
-            reshaped_base_tensor = aliased_base_tensor
-        out = target_meta_tensor._view_func(reshaped_base_tensor)
-        # This shape mismatch can happen due to a bug in inplace/view handling in autograd.
-        # Try putting a breakpoint here and running
-        # `test/functorch/test_aotdispatch TestAOTAutograd.test_output_all_alias_types`
-        # Also, https://github.com/pytorch/pytorch/issues/49825
-        #
-        # As a stopgap, we'll fall back to as_strided.
-        if out is not None and out.shape == target_meta_tensor.shape:
-            if aliased_base_tensor.requires_grad and not target_requires_grad:
-                out = out.detach()
-            elif not aliased_base_tensor.requires_grad and target_requires_grad:
-                out.requires_grad_(True)
-            return out
-    size = target_meta_tensor.size()
-    stride = target_meta_tensor.stride()
-    storage_offset = target_meta_tensor.storage_offset()
-    if aliased_base_tensor.is_complex() and not target_meta_tensor.is_complex():
-        aliased_out = torch.view_as_real(aliased_base_tensor).as_strided(
-            size, stride, storage_offset
-        )
-    elif not aliased_base_tensor.is_complex() and target_meta_tensor.is_complex():
-        aliased_out = torch.view_as_complex(aliased_base_tensor).as_strided(
-            size, stride, storage_offset
-        )
-    else:
-        aliased_out = aliased_base_tensor.as_strided(size, stride, storage_offset)
-    # For outputs aliasing inputs, we need to check if the requires-gradness has changed.
-    if aliased_base_tensor.requires_grad and not target_requires_grad:
-        aliased_out = aliased_out.detach()
-    elif not aliased_base_tensor.requires_grad and target_requires_grad:
-        aliased_out.requires_grad_(True)
-    # For outputs aliasing inputs, we need to check if the dtype has changed.
-    # as_strided() is the "most generic" view, but it does not cover cross-dtype views
-    if aliased_out.dtype != target_meta_tensor.dtype:
-        aliased_out = aliased_out.view(target_meta_tensor.dtype)
-    return aliased_out
-
-def to_fun(t):
-    if isinstance(t, Tensor):
-        if is_traceable_wrapper_subclass(t):
-            # See Note [Functionalization always runs last]
-            # This means that if we want to "functionalize" a subclass, we need to ensure that the functional wrapper
-            # goes at the bottom.
-            # recurse here, so we can support nested wrapper subclasses
-            out = transform_subclass(t, lambda _, inner_t: to_fun(inner_t))
-            torch._mirror_autograd_meta_to(t, out)
-            return out
-        else:
-            return FunctionalTensor.to_functional(t)
-    else:
-        return t
-
-def sync_functional_tensor(t):
-    if is_traceable_wrapper_subclass(t):
-        attrs, ctx = t.__tensor_flatten__()
-        for attr in attrs:
-            sync_functional_tensor(getattr(t, attr))
-    else:
-        torch._sync(t)
-
-# When subclasses are involved, t here will usually look something like:
-# SubclassA(SubclassB(FunctionalTensor(_to_functional_tensor(FakeTensor))))
-def from_fun(t):
-    if isinstance(t, Tensor) and is_traceable_wrapper_subclass(t):
-        # See Note [Functionalization always runs last]
-        # This means that if we want to "functionalize" a subclass, we need to ensure that the functional wrapper
-        # goes at the bottom.
-        # recurse here, so we can support nested wrapper subclasses
-        out = transform_subclass(t, lambda _, inner_t: from_fun(inner_t))
-        torch._mirror_autograd_meta_to(t, out)
-        return out
-
-    if not isinstance(t, FunctionalTensor):
-        # quick sanity assert
-        if isinstance(t, torch.Tensor):
-            assert not torch._is_functional_tensor(t)
-        return t
-    sync_functional_tensor(t)
-    return torch._from_functional_tensor(t.elem)
-
-def is_fun(t):
-    if isinstance(t, Tensor) and is_traceable_wrapper_subclass(t):
-        # See Note [Functionalization always runs last]
-        # This means that if we want to "functionalize" a subclass, we need to ensure that the functional wrapper
-        # goes at the bottom.
-        # recurse here, so we can support nested wrapper subclasses
-        t_attrs, _ = t.__tensor_flatten__()
-        t_inners = [getattr(t, attr) for attr in t_attrs]
-        any_fun = any(is_fun(x) for x in t_inners)
-        all_fun = all(is_fun(x) for x in t_inners)
-        assert any_fun == all_fun
-        return any_fun
-
-    return isinstance(t, FunctionalTensor)
-
-# t here is either
-# (1) A FunctionalTensor(_to_functional_tensor(FakeTensor))
-# (2) A traceable tensor subclass that holds a FunctionalTensor
-def has_metadata_mutation(t):
-    if is_traceable_wrapper_subclass(t):
-        attrs, _ = t.__tensor_flatten__()
-        # A tensor subclass was updated if any of its inner elements were updated
-        return any(has_metadata_mutation(getattr(t, attr)) for attr in attrs)
-    else:
-        assert isinstance(t, FunctionalTensor)
-        return torch._functionalize_has_metadata_mutation(t.elem)
-
-def are_all_mutations_hidden_from_autograd(t):
-    if is_traceable_wrapper_subclass(t):
-        attrs, _ = t.__tensor_flatten__()
-        # If all inner elements are mutations hidden from autograd, then it is a mutation hidden from autograd.
-        return all(are_all_mutations_hidden_from_autograd(getattr(t, attr)) for attr in attrs)
-    else:
-        assert isinstance(t, FunctionalTensor)
-        return torch._functionalize_are_all_mutations_hidden_from_autograd(t.elem)
-
-# new_arg and arg here are either:
-# (1) both a FakeTensor
-# (2) both a traceable tensor subclass that holds a FakeTensor
-# Pre-condition: the two args are the "old" and "new" inputs from running functionalization.
-# When we run functionalization and wrap our inputs into FunctionalTensors,
-# we can detect whether or not an input was mutated by checking to see if the inner tensor has changed
-#
-# Normally it would be enough just to check if arg is new_arg, which is normally enough for functionalization
-# to confirm that inputs were not mutated when running the user's model with functionalization on.
-# But when we have subclass inputs, we can't rely on that:
-# `from_fun(to_fun(x)) is x` will return False, because the call to `from_fun` constructs
-# a brand new subclass instance: we are calling __tensor_unflatten__, and going
-# from Subclass(FakeTensor) to Subclass(FunctionalTensor(FakeTensor))
-def was_updated(arg, new_arg):
-    if is_traceable_wrapper_subclass(arg):
-        assert is_traceable_wrapper_subclass(new_arg)
-        attrs, _ = arg.__tensor_flatten__()
-        new_attrs, _ = new_arg.__tensor_flatten__()
-        assert attrs == new_attrs
-        # A tensor subclass was updated if any of its inner elements were updated
-        return any(was_updated(getattr(arg, attr), getattr(new_arg, attr)) for attr in attrs)
-    else:
-        return arg is not new_arg
-
-# new_arg and arg here are either:
-# (1) both a FakeTensor
-# (2) both a traceable tensor subclass that holds a FakeTensor
-# Pre-condition: the two args are the "old" and "new" inputs from running functionalization.
-# When we run functionalization and wrap our inputs into FunctionalTensors,
-# we can detect whether or not an input was mutated by checking to see if the inner tensor has changed,
-# but shares storage with the old input
-def was_metadata_updated(arg, new_arg):
-    if is_traceable_wrapper_subclass(arg):
-        assert is_traceable_wrapper_subclass(new_arg)
-        attrs, _ = arg.__tensor_flatten__()
-        new_attrs, _ = new_arg.__tensor_flatten__()
-        assert attrs == new_attrs
-        # A tensor subclass was updated if any of its inner elements were updated
-        return any(was_metadata_updated(getattr(arg, attr), getattr(new_arg, attr)) for attr in attrs)
-    else:
-        return arg is not new_arg and StorageWeakRef(arg.untyped_storage()) == StorageWeakRef(new_arg.untyped_storage())
 
 def requires_subclass_dispatch(args, fw_metadata: ViewAndMutationMeta) -> bool:
     args_flattened = pytree.arg_tree_leaves(*args)
@@ -953,11 +779,11 @@ def run_functionalized_fw_and_collect_metadata(
 ) -> ViewAndMutationMeta:
     memo = {}
 
-    def _to_fun(t):
+    def _to_functional(t):
         if isinstance(t, Tensor):
             if t in memo:
                 return memo[t]
-            r = to_fun(t)
+            r = to_functional(t)
             memo[t] = r
             return r
         else:
@@ -971,7 +797,7 @@ def run_functionalized_fw_and_collect_metadata(
         input_info: List[InputAliasInfo] = []
         output_info: List[OutputAliasInfo] = []
 
-        flat_f_args = pytree.tree_map(_to_fun, flat_args)
+        flat_f_args = pytree.tree_map(_to_functional, flat_args)
 
         prior_grad_enabled = torch.is_grad_enabled()
         prior_autocast_states = _get_autocast_states()
@@ -996,9 +822,9 @@ def run_functionalized_fw_and_collect_metadata(
             if not isinstance(arg, Tensor):
                 new_arg = arg
             else:
-                new_arg = from_fun(f_arg)
-            if was_updated(arg, new_arg):
-                if was_metadata_updated(arg, new_arg):
+                new_arg = from_functional(f_arg)
+            if was_tensor_updated(arg, new_arg):
+                if was_tensor_metadata_updated(arg, new_arg):
                     mutates_data = False
                     mutates_metadata = True
                 else:
@@ -1309,9 +1135,9 @@ from a multi-output view call")
         ]
         # intermediate bases are also included in the backward graph
         f_tangents = f_input_tangents + f_output_tangents + intermediate_bases
-        traced_tangents = pytree.tree_map(from_fun, f_tangents)
+        traced_tangents = pytree.tree_map(from_functional, f_tangents)
         traced_tangents = pytree.tree_map(view_avoid_dupes_with_primals, traced_tangents)
-        user_outs = pytree.tree_map(from_fun, f_output_tangents)
+        user_outs = pytree.tree_map(from_functional, f_output_tangents)
 
         f_mutated_inputs = [
             inp
@@ -1339,7 +1165,7 @@ from a multi-output view call")
             f_fw_graph_outs = f_metadata_mutated_inputs + f_fw_graph_outs
         if is_train:
             f_fw_graph_outs = f_fw_graph_outs + intermediate_bases
-        fw_graph_outs = pytree.tree_map(from_fun, f_fw_graph_outs)
+        fw_graph_outs = pytree.tree_map(from_functional, f_fw_graph_outs)
 
         grad_enabled_mutation = None
         if torch.is_grad_enabled() != prior_grad_enabled:
@@ -1760,7 +1586,7 @@ def create_functionalized_fn(
 ) -> Tuple[Callable, List[Any]]:
     def functionalized_f_helper(*args):
         # Wrap inputs into functional wrappers
-        f_args = pytree.tree_map(to_fun, args)
+        f_args = pytree.tree_map(to_functional, args)
 
         # See Note [Disabling Functionalize TLS Above Python Functionalization]
         disable_above = torch._C._ExcludeDispatchKeyGuard(torch._C.DispatchKeySet(torch._C.DispatchKey.Functionalize))
@@ -1796,8 +1622,8 @@ def create_functionalized_fn(
             for i, (inpt_old, inpt_f) in enumerate(zip(args, f_args) if not trace_joint else zip(args[0], f_args[0])):
                 if not isinstance(inpt_f, torch.Tensor):
                     continue
-                assert is_fun(inpt_f)
-                inpt_new = from_fun(inpt_f)
+                assert is_functional(inpt_f)
+                inpt_new = from_functional(inpt_f)
                 if meta.input_info[i].mutation_type == MutationType.MUTATED_IN_GRAPH:
                     # We found an input that had a (data-only) mutation.
                     # Since keep_input_mutations is set, we need to faithfully apply a copy_()
@@ -1808,7 +1634,7 @@ def create_functionalized_fn(
                     else:
                         inpt_old.copy_(inpt_new)
 
-        return pytree.tree_map(from_fun, f_outs)
+        return pytree.tree_map(from_functional, f_outs)
 
     # Kinda annoying, but needed to make sure that the fx graph we trace out has "primals"
     # and "tangents" as its input names (which are special-cased by the partitioner)
